Ignore .vscode files and modify setup.sh to check for existance of env directory

+++ setup.sh
@@ -1,2 +1,7 @@
 #! /bin/bash
-python3 -m venv env; source env/bin/activate; python -m pip install -r requirements.txt
+if [ ! -d 'env' ]; then # if env directory does NOT already exist then set it up
+	python3 -m venv env; source env/bin/activate; python -m pip install -r requirements.txt
+else # else inform user env already exists
+	echo "Virtual environment already exists"
+	echo "activate with 'source env/bin/activate'"
+fi
